fix: audit findings, including split-child geometry aliasing (#33)
Security audit of the post-refactor tree. No critical or high findings;
these are the medium and low ones, plus a coverage gap it uncovered.

M-1, consensus-critical. sampleSplitChild salted its stream with
uint8(childIndex) while _childSeed uses the full index, so children 256
apart in one split shared a sampling stream and stored byte-identical
geometry while holding different ids and seeds. A 100 ETH parent split
into dust hits this repeatedly. The stream now takes the untruncated
index, applied in lockstep across the library, the TypeScript canonical
mirror, SAMPLING_SPEC section 6 (where the oversight originated), and
regenerated parity fixtures. genFixtures gains a paired 0-vs-256 vector
so parity pins non-aliasing on both sides.

L-1. bid checked settled, then handed control to the Shapes fee
recipient through the escrow mint. A fee recipient that is also the
seller could cancel mid-call and have the bid written onto a settled
auction. bid now re-checks settlement after the escrow interaction,
which closes the class without giving settle and cancelAuction the
reentrancy guard they deliberately omit.

L-2. _burnForRedemption cleared _shapes but not _sampledModules,
unlike the compose and split paths. Dead state and a missed refund
rather than a live bug, since no path revives a redeemed id, but the
asymmetry was a trap. sacrifice is deliberately left alone: it does not
burn, and the Black token still renders from those bytes.

I-1. The renderer's canonical draw-order docstring described a wRatio
stream draw that the code does not make (WRATIO is a constant) and
numbered the per-cell draws wrongly. Corrected against the code.

Adds the missing decompose test: a materialized burned input now has to
round-trip bit-exactly, which was the one part of the decompose promise
the suite never asserted.

415 tests pass (+4), parity 11/11 against regenerated fixtures, sizes
gate exits 0. Shapes runtime margin narrows 174 to 135 bytes.

+++ src/lib/GeometrySampling.sol
@@ -160,6 +160,10 @@ library GeometrySampling {
     /// @notice Sample one split child's module array from its parent's effective modules
     ///         (SAMPLING_SPEC.md §6). Uniform over the parent's modules, with replacement: a
     ///         single donor, so no units weighting.
+    /// @dev `childIndex` enters the stream as a full uint256, matching the untruncated index
+    ///      `Shapes._childSeed` uses. Encoding it as a uint8 would give children at index i and
+    ///      i + 256 of the same parent at the same denomination an identical stream, and so
+    ///      identical stored modules.
     function sampleSplitChild(
         bytes memory parentModules,
         bytes32 parentSeed,
@@ -167,7 +171,7 @@ library GeometrySampling {
         uint256 childIndex
     ) public pure returns (bytes memory result) {
         Round03Rand.Stream memory rnd = Round03Rand.init(
-            keccak256(abi.encodePacked("Shapes/sample-split/v1", parentSeed, childDenom, uint8(childIndex)))
+            keccak256(abi.encodePacked("Shapes/sample-split/v1", parentSeed, childDenom, childIndex))
         );
 
         (uint256 cols, uint256 rows) = Denominations.gridAt(childDenom);
